% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_netmhc2pan_bin_url.R
\name{check_netMHCpan_bin_url}
\alias{check_netMHCpan_bin_url}
\title{Check the URL of the NetMHCIIpan binary tarball}
\usage{
check_netMHCpan_bin_url(
  netMHCpan_bin_url = get_netMHCpan_bin_url(),
  verbose = FALSE,
  netMHCpan_archive_filename = get_netMHCpan_archive_filename(),
  temp_local_file = tempfile(pattern = "check_netMHCpan_bin_url_")
)
}
\arguments{
\item{netMHCpan_bin_url}{URL to download the
NetMHCpan binary tarball file from,
similar to, for example,
\code{https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/data.tar.gz}.
Use \link{get_netMHCpan_bin_url} to get the default URL.}

\item{verbose}{set to TRUE for more output}

\item{netMHCpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netMHCpan_archive_filename} to get this filename.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\description{
Check the URL of the NetMHCIIpan binary tarball is valid,
will stop if not.
This URL link expires after 4 hours.
}
