% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_netmhcpan_installation.R
\name{check_netmhcpan_installation}
\alias{check_netmhcpan_installation}
\title{Checks the installation of NetMHCpan.}
\usage{
check_netmhcpan_installation(
  netmhcpan_folder_name = get_default_netmhcpan_folder()
)
}
\arguments{
\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\value{
Nothing
}
\description{
Checks the installation of NetMHCpan.
Throws a helpful error message if incomplete,
else does nothing
}
\examples{
# Will stop if NetMHCpan is not installed
try(check_netmhcpan_installation())
}
\author{
Richèl J.C. Bilderbeek
}
