% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_fasta_filename.R
\name{create_temp_fasta_filename}
\alias{create_temp_fasta_filename}
\title{Create a \code{.fasta} file}
\usage{
create_temp_fasta_filename()
}
\value{
a path to a non-existing file,
for example, \code{/home/myusername/.cache/temp_582046426735.fasta}
}
\description{
Create a \code{.fasta} file
}
\examples{
create_temp_fasta_filename()
}
\author{
Richèl J.C. Bilderbeek
}
