% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_xls_filename.R
\name{create_temp_xls_filename}
\alias{create_temp_xls_filename}
\title{Create a \code{.xls} file}
\usage{
create_temp_xls_filename()
}
\value{
a path to a non-existing file,
for example, \code{/home/myusername/.cache/temp_582047dac733.xls}
}
\description{
Create a \code{.xls} file
}
\examples{
create_temp_xls_filename()
}
\author{
Richèl J.C. Bilderbeek
}
