% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_netmhcpan_bin.R
\name{download_netmhcpan_bin}
\alias{download_netmhcpan_bin}
\title{Download the NetMHCpan binary}
\usage{
download_netmhcpan_bin(
  netmhcpan_bin_url = get_netmhcpan_bin_url(),
  netmhcpan_bin_tarfile_path = get_default_netmhcpan_bin_tarfile_path(),
  verbose = FALSE,
  netmhcpan_archive_filename = get_netmhcpan_archive_filename(),
  temp_local_file = tempfile(pattern = "netmhcpan_download_netmhcpan_")
)
}
\arguments{
\item{netmhcpan_bin_url}{URL to download the
NetMHCpan binary tarball file from,
similar to, for example,
\code{ FIX LINK}.
Use \link{get_netmhcpan_bin_url} to get the default URL.}

\item{netmhcpan_bin_tarfile_path}{path of the
NetMHCpan binary tarball file}

\item{verbose}{set to TRUE for more output}

\item{netmhcpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1b.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netmhcpan_archive_filename} to get this filename.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\value{
Nothing
}
\description{
Download the NetMHCpan binary tarball file
}
\examples{
\donttest{
  install_netmhcpan_bin()
}
}
\author{
Richèl J.C. Bilderbeek
}
