% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhcpan_bin_path.R
\name{get_default_netmhcpan_bin_path}
\alias{get_default_netmhcpan_bin_path}
\title{Get the full path to the default NetMHCpan binary}
\usage{
get_default_netmhcpan_bin_path(
  netmhcpan_subfolder = get_default_netmhcpan_subfolder()
)
}
\arguments{
\item{netmhcpan_subfolder}{the subfolder (to be) used by NetMHCpan.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder}
}
\value{
the full path to the default NetMHCpan binary
}
\description{
Get the full path to the default NetMHCpan binary
}
\examples{
get_default_netmhcpan_bin_path()
}
\author{
Richèl J.C. Bilderbeek
}
