% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhc2pan_subfolder.R
\name{get_default_netMHCpan_subfolder}
\alias{get_default_netMHCpan_subfolder}
\title{Get the full path to the NetMHCpan sub-folder}
\usage{
get_default_netMHCpan_subfolder(
  netMHCpan_folder_name = get_default_netMHCpan_folder(),
  netMHCpan_version = get_default_netMHCpan_version()
)
}
\arguments{
\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{netMHCpan_version}{the NetMHCpan version,
for example \code{4.1}.
Use \link{get_default_netMHCpan_version} to get the default
NetMHCpan version}
}
\value{
the full path to the NetMHCpan sub-folder
}
\description{
Get the full path to the NetMHCpan sub-folder
}
\examples{
get_default_netMHCpan_subfolder()
}
\author{
Richèl J.C. Bilderbeek
}
