% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhc2pan_alleles.R
\name{get_netMHCpan_alleles}
\alias{get_netMHCpan_alleles}
\title{Get a list of the alleles supported by NetMHCIIpan}
\usage{
get_netMHCpan_alleles(netMHCpan_folder_name = get_default_netMHCpan_folder())
}
\arguments{
\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\value{
a character vector with the NetMHCIIpan alleles
}
\description{
Get a list of the alleles supported by NetMHCIIpan
}
\examples{
if (is_netmhc2pan_installed()) {
  get_netmhc2pan_alleles()[1:5]
}
}
\author{
Richèl J.C. Bilderbeek
}
