% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhc2pan_archive_filename.R
\name{get_netMHCpan_archive_filename}
\alias{get_netMHCpan_archive_filename}
\title{Get the filename of the NetMHCpan archive file}
\usage{
get_netMHCpan_archive_filename(
  netMHCpan_version = get_default_netMHCpan_version()
)
}
\arguments{
\item{netMHCpan_version}{the NetMHCpan version,
for example \code{4.1}.
Use \link{get_default_netMHCpan_version} to get the default
NetMHCpan version}
}
\value{
the filename of the NetMHCpan archive file
}
\description{
Get the filename of the NetMHCpan archive file
}
\examples{
get_netMHCpan_archive_filename()
}
