% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhcpan_url.R
\name{get_netmhcpan_url}
\alias{get_netmhcpan_url}
\title{Deprecated, use \link{get_netmhcpan_bin_url} instead}
\usage{
get_netmhcpan_url()
}
\value{
a download URL
}
\description{
Deprecated, use \link{get_netmhcpan_bin_url} instead
}
\author{
Richèl J.C. Bilderbeek
}
