% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhcpan.R
\name{install_netmhcpan}
\alias{install_netmhcpan}
\title{Install NetMHCpan}
\usage{
install_netmhcpan(
  netmhcpan_bin_url = get_netmhcpan_bin_url(),
  netmhcpan_data_url = get_netmhcpan_data_url(),
  verbose = FALSE,
  netmhcpan_archive_filename = get_netmhcpan_archive_filename(),
  netmhcpan_folder_name = get_default_netmhcpan_folder(),
  temp_local_file = tempfile(pattern = "netmhcpan_install_netmhcpan_")
)
}
\arguments{
\item{netmhcpan_bin_url}{URL to download the
NetMHCpan binary tarball file from,
similar to, for example,
\code{ FIX LINK}.
Use \link{get_netmhcpan_bin_url} to get the default URL.}

\item{netmhcpan_data_url}{URL to download the
NetMHCpan data tarball file from,
similar to, for example,
\url{https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/data.tar.gz}.
Use \link{get_netmhcpan_data_url} to get the default URL.}

\item{verbose}{set to TRUE for more output}

\item{netmhcpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1b.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netmhcpan_archive_filename} to get this filename.}

\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\value{
Nothing
}
\description{
Install NetMHCpan to a local folder, by downloading the binary and
data files. Use \link{install_netmhcpan_from_files} to install
NetMHCpan from files that are already downloaded.
}
\details{
These are three steps:

\itemize{
  \item Install the NetMHCpan binary, using \link{install_netmhcpan_bin}
  \item Install the NetMHCpan data, using \link{install_netmhcpan_data}
  \item Set up NetMHCpan, using \link{set_up_netmhcpan}
}
}
\examples{
\donttest{
  install_netmhcpan()
}
}
\author{
Richèl J.C. Bilderbeek
}
