% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhcpan_bin_from_file.R
\name{install_netmhcpan_bin_from_file}
\alias{install_netmhcpan_bin_from_file}
\title{Install the NetMHCpan binary to a local folder}
\usage{
install_netmhcpan_bin_from_file(
  netmhcpan_bin_tarfile_path,
  verbose = FALSE,
  netmhcpan_archive_filename = get_netmhcpan_archive_filename(),
  netmhcpan_folder_name = get_default_netmhcpan_folder()
)
}
\arguments{
\item{netmhcpan_bin_tarfile_path}{path of the
NetMHCpan binary tarball file}

\item{verbose}{set to TRUE for more output}

\item{netmhcpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1b.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netmhcpan_archive_filename} to get this filename.}

\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\value{
Nothing
}
\description{
Install the NetMHCpan binary to a local folder
}
\author{
Richèl J.C. Bilderbeek
}
