% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhcpan_data_from_file.R
\name{install_netmhcpan_data_from_file}
\alias{install_netmhcpan_data_from_file}
\title{Install the NetMHCpan data from file}
\usage{
install_netmhcpan_data_from_file(
  netmhcpan_data_tarfile_path,
  netmhcpan_folder_name = rappdirs::user_data_dir(),
  verbose = FALSE
)
}
\arguments{
\item{netmhcpan_data_tarfile_path}{path of the
NetMHCpan data tarball file}

\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\value{
Nothing
}
\description{
Install the NetMHCpan data from the NetMHCpan data tarball file
to the default NetMHCpan folder.
}
\details{
The NetMHCpan data tarball file can be in any path.

The data tarball file can be downloaded from
\url{https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/data.tar.gz},
without filling in a contact form.
}
\author{
Richèl J.C. Bilderbeek
}
