% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhc2pan_installed.R
\name{is_netMHCpan_installed}
\alias{is_netMHCpan_installed}
\title{Measure if NetMHCpan is installed locally}
\usage{
is_netMHCpan_installed(
  netMHCpan_folder_name = get_default_netMHCpan_folder(),
  verbose = FALSE
)
}
\value{
TRUE is NetMHCpan is installed locally,
  FALSE otherwise
}
\description{
Measure if NetMHCpan is installed locally
}
\author{
Richèl J.C. Bilderbeek
}
