% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhcpan_installed.R
\name{is_netmhcpan_set_up}
\alias{is_netmhcpan_set_up}
\title{Measure if NetMHCpan is set up}
\usage{
is_netmhcpan_set_up(netmhcpan_folder_name = get_default_netmhcpan_folder())
}
\arguments{
\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\value{
TRUE is NetMHCpan is set up locally,
  FALSE otherwise
}
\description{
Measure if NetMHCpan is set up
}
\author{
Richèl J.C. Bilderbeek
}
