% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url_valid.R
\name{is_url_valid}
\alias{is_url_valid}
\title{Is the download URL valid?}
\usage{
is_url_valid(
  netmhcpan_bin_url = get_netmhcpan_bin_url(),
  verbose = FALSE,
  netmhcpan_archive_filename = get_netmhcpan_archive_filename(),
  temp_local_file = tempfile(pattern = "netmhcpan_is_url_valid_")
)
}
\arguments{
\item{netmhcpan_bin_url}{URL to download the
NetMHCpan binary tarball file from,
similar to, for example,
\code{ FIX LINK}.
Use \link{get_netmhcpan_bin_url} to get the default URL.}

\item{verbose}{set to TRUE for more output}

\item{netmhcpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1b.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netmhcpan_archive_filename} to get this filename.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\description{
The download link expires after 4 hours.
}
