% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan_report.R
\name{netmhcpan_report}
\alias{netmhcpan_report}
\title{Create a \code{netmhcpan} report, to be used when reporting bugs}
\usage{
netmhcpan_report(netmhcpan_folder_name = get_default_netmhcpan_folder())
}
\arguments{
\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\description{
Create a \code{netmhcpan} report, to be used when reporting bugs
}
\examples{
netmhcpan_report()
}
\author{
Richèl J.C. Bilderbeek
}
