% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmhcpan_self_test.R
\name{netmhcpan_self_test}
\alias{netmhcpan_self_test}
\title{Test NetMHCpan}
\usage{
netmhcpan_self_test(
  netmhcpan_folder_name = get_default_netmhcpan_folder(),
  verbose = FALSE
)
}
\arguments{
\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\description{
Test NetMHCpan
}
\author{
Richèl J.C. Bilderbeek
}
