% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_netmhc2pan.R
\name{run_netMHCpan}
\alias{run_netMHCpan}
\title{Run NetMHCpan}
\usage{
run_netMHCpan(
  fasta_filename,
  alleles = "DRB1_0101",
  peptide_length = 15,
  netMHCpan_folder_name = get_default_netMHCpan_folder(),
  temp_xls_filename = netMHCpan::create_temp_xls_filename()
)
}
\arguments{
\item{fasta_filename}{the name of a FASTA file with protein sequences}

\item{alleles}{one or more alleles, e.g. \code{DRB1_0101}.
See \link{get_netMHCpan_alleles} for a full list}

\item{peptide_length}{length of a peptide}

\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{temp_xls_filename}{name for a temporary \code{xls} file,
which will be deleted automatically}
}
\value{
a data frame with the NetMHCpan results
}
\description{
Run NetMHCpan
}
\examples{
if (is_netMHCpan_installed()) {

  fasta_filename <- system.file(
    "extdata", "example.fasta", package = "netMHCpan"
  )
  run_netMHCpan(fasta_filename)

  # Two alleles
  alleles <- c("DRB1_0101", "DRB1_0102")
  # Run NetMHCpan with these two alleles
  run_netmhc2pan(fasta_filename, alleles = alleles)
}
}
\author{
Richèl J.C. Bilderbeek
}
