% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_netmhc2pan.R
\name{set_up_netMHCpan}
\alias{set_up_netMHCpan}
\title{Install the NetMHCpan binary to a local folder}
\usage{
set_up_netMHCpan(
  netMHCpan_folder_name = get_default_netMHCpan_folder(),
  verbose = FALSE
)
}
\arguments{
\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\value{
Nothing
}
\description{
Install the NetMHCpan binary to a local folder
}
\examples{
\donttest{
  set_up_netMHCpan()
}
}
\author{
Richèl J.C. Bilderbeek
}
