% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_netmhcpan_name.R
\name{to_netmhcpan_name}
\alias{to_netmhcpan_name}
\title{Convert a formal haplotype name to its NetMHCpan notation}
\usage{
to_netmhcpan_name(mhc_haplotype)
}
\arguments{
\item{mhc_haplotype}{one MHC haplotype, e.g. \code{HLA-A0101}.
See \link{get_netmhcpan_alleles} for a full list}
}
\description{
Convert a formal MHC haplotype name to the notation used by NetMHCpan.
}
