% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_netmhc2pan.R
\name{uninstall_netMHCpan}
\alias{uninstall_netMHCpan}
\title{Uninstall NetMHCpan}
\usage{
uninstall_netMHCpan(netMHCpan_folder_name = get_default_netMHCpan_folder())
}
\arguments{
\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\description{
Uninstall NetMHCpan
}
\author{
Richèl J.C. Bilderbeek
}
