% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_netmhcpan_bin_url.R
\name{check_netmhcpan_bin_url}
\alias{check_netmhcpan_bin_url}
\title{Check the URL of the NetMHCpan binary tarball}
\usage{
check_netmhcpan_bin_url(
  netmhcpan_bin_url = get_netmhcpan_bin_url(),
  verbose = FALSE,
  netmhcpan_archive_filename = get_netmhcpan_archive_filename(),
  temp_local_file = tempfile(pattern = "check_netmhcpan_bin_url_")
)
}
\arguments{
\item{netmhcpan_bin_url}{URL to download the
NetMHCpan binary tarball file from,
similar to, for example,
\code{ FIX LINK}.
Use \link{get_netmhcpan_bin_url} to get the default URL.}

\item{verbose}{set to TRUE for more output}

\item{netmhcpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1b.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netmhcpan_archive_filename} to get this filename.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\description{
Check the URL of the NetMHCpan binary tarball is valid,
will stop if not.
This URL link expires after 4 hours.
}
