% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_netmhc2pan_installation.R
\name{check_netMHCpan_installation}
\alias{check_netMHCpan_installation}
\title{Checks the installation of NetMHCIIpan.}
\usage{
check_netMHCpan_installation(
  netMHCpan_folder_name = get_default_netMHCpan_folder()
)
}
\arguments{
\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\value{
Nothing
}
\description{
Checks the installation of NetMHCIIpan.
Throws a helpful error message if incomplete,
else does nothing
}
\examples{
# Will stop if NetMHC2pan is not installed
try(check_netmhc2pan_installation())
}
\author{
Richèl J.C. Bilderbeek
}
