% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  alleles,
  do_filter,
  fasta_filename,
  folder_name,
  mhc_haplotype,
  netMHCpan_archive_filename,
  netMHCpan_bin_tarfile_path,
  netMHCpan_bin_url,
  netMHCpan_data_tarfile_path,
  netMHCpan_data_url,
  netMHCpan_folder_name,
  netMHCpan_subfolder,
  netMHCpan_version,
  os,
  peptide_length,
  peptides,
  protein_sequence,
  temp_fasta_filename,
  temp_local_file,
  temp_xls_filename,
  verbose,
  xls_filename
)
}
\arguments{
\item{alleles}{one or more alleles, e.g. \code{DRB1_0101}.
See \link{get_netMHCpan_alleles} for a full list}

\item{do_filter}{set to TRUE if the results of NetMHCpan must be filtered}

\item{fasta_filename}{the name of a FASTA file with protein sequences}

\item{folder_name}{the folder to install NetMHCpan, which is
\code{"/home/[usename]/.local/share"} by default}

\item{mhc_haplotype}{one MHC haplotype, e.g. \code{DRB1_0101}.
See \link{get_netMHCpan_alleles} for a full list}

\item{netMHCpan_archive_filename}{the NetMHCpan archive
filename, for example \code{NetMHCpan-4.1/data.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netMHCpan_archive_filename} to get this filename.}

\item{netMHCpan_bin_tarfile_path}{path of the
NetMHCpan binary tarball file}

\item{netMHCpan_bin_url}{URL to download the
NetMHCpan binary tarball file from,
similar to, for example,
\code{https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/data.tar.gz}.
Use \link{get_netMHCpan_bin_url} to get the default URL.}

\item{netMHCpan_data_tarfile_path}{path of the
NetMHCpan data tarball file}

\item{netMHCpan_data_url}{URL to download the
NetMHCpan data tarball file from,
similar to, for example,
\url{https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/data.tar.gz}.
Use \link{get_netMHCpan_data_url} to get the default URL.}

\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{netMHCpan_subfolder}{the subfolder (to be) used by NetMHCpan.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder}

\item{netMHCpan_version}{the NetMHCpan version,
for example \code{4.1}.
Use \link{get_default_netMHCpan_version} to get the default
NetMHCpan version}

\item{os}{the operating system as obtained by \code{rappdirs::app_dir()$os}.
\code{netMHCpan} supports Linux ('\code{unix}') only}

\item{peptide_length}{length of a peptide}

\item{peptides}{one or more peptide sequences}

\item{protein_sequence}{a protein sequence, for example \code{FAMILYVW}}

\item{temp_fasta_filename}{name for a temporary FASTA file,
which will be deleted automatically}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}

\item{temp_xls_filename}{name for a temporary \code{xls} file,
which will be deleted automatically}

\item{verbose}{set to TRUE for more output}

\item{xls_filename}{name of an \code{xls} file}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
