% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_netmhc2pan_bin.R
\name{download_netMHCpan_bin}
\alias{download_netMHCpan_bin}
\title{Download the NetMHCpan binary}
\usage{
download_netMHCpan_bin(
  netMHCpan_bin_url = get_netMHCpan_bin_url(),
  netMHCpan_bin_tarfile_path = get_default_netMHCpan_bin_tarfile_path(),
  verbose = FALSE,
  netMHCpan_archive_filename = get_netMHCpan_archive_filename(),
  temp_local_file = tempfile(pattern = "netMHCpan_download_netMHCpan_")
)
}
\arguments{
\item{netMHCpan_bin_url}{URL to download the
NetMHCpan binary tarball file from,
similar to, for example,
\code{https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/data.tar.gz}.
Use \link{get_netMHCpan_bin_url} to get the default URL.}

\item{netMHCpan_bin_tarfile_path}{path of the
  NetMHCpan binary tarball file
##This needs CHANGING}

\item{verbose}{set to TRUE for more output}

\item{netMHCpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netMHCpan_archive_filename} to get this filename.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\value{
Nothing
}
\description{
Download the NetMHCpan binary tarball file
}
\examples{
\donttest{
  install_netMHCpan_bin()
}
}
\author{
Richèl J.C. Bilderbeek
}
