% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhc2pan_bin_path.R
\name{get_default_netMHCpan_bin_path}
\alias{get_default_netMHCpan_bin_path}
\title{Get the full path to the default NetMHCpan binary}
\usage{
get_default_netMHCpan_bin_path(
  netMHCpan_subfolder = get_default_netMHCpan_subfolder()
)
}
\arguments{
\item{netMHCpan_subfolder}{the subfolder (to be) used by NetMHCpan.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder}
}
\value{
the full path to the default NetMHCpan binary
}
\description{
Get the full path to the default NetMHCpan binary
}
\examples{
get_default_netMHCpan_bin_path()
}
\author{
Richèl J.C. Bilderbeek
}
