% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhcpan_bin_tarfile_path.R
\name{get_default_netmhcpan_bin_tarfile_path}
\alias{get_default_netmhcpan_bin_tarfile_path}
\title{Get the default path for the NetMHCpan binary tarball file}
\usage{
get_default_netmhcpan_bin_tarfile_path(
  netmhcpan_folder_name = get_default_netmhcpan_folder(),
  netmhcpan_archive_filename = get_netmhcpan_archive_filename()
)
}
\arguments{
\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{netmhcpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1b.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netmhcpan_archive_filename} to get this filename.}
}
\description{
Get the default path for the NetMHCpan binary tarball file.
This is the location where it will be saved to after downloading.
}
\examples{
get_default_netmhcpan_bin_tarfile_path()
}
\author{
Richèl J.C. Bilderbeek
}
