% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhcpan_subfolder.R
\name{get_default_netmhcpan_subfolder}
\alias{get_default_netmhcpan_subfolder}
\title{Get the full path to the NetMHCpan sub-folder}
\usage{
get_default_netmhcpan_subfolder(
  netmhcpan_folder_name = get_default_netmhcpan_folder(),
  netmhcpan_version = get_default_netmhcpan_version()
)
}
\arguments{
\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{netmhcpan_version}{the NetMHCpan version,
for example \code{4.1}.
Use \link{get_default_netmhcpan_version} to get the default
NetMHCpan version}
}
\value{
the full path to the NetMHCpan sub-folder
}
\description{
Get the full path to the NetMHCpan sub-folder
}
\examples{
get_default_netmhcpan_subfolder()
}
\author{
Richèl J.C. Bilderbeek
}
