% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhcpan_version.R
\name{get_default_netmhcpan_version}
\alias{get_default_netmhcpan_version}
\title{Get the default NetMHCpan version used}
\usage{
get_default_netmhcpan_version()
}
\value{
the default NetMHCpan version used
}
\description{
Get the default NetMHCpan version used
}
\examples{
get_default_netmhcpan_version()
}
\author{
Richèl J.C. Bilderbeek
}
