% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhcpan_alleles.R
\name{get_netmhcpan_alleles}
\alias{get_netmhcpan_alleles}
\title{Get a list of the alleles supported by NetMHCpan}
\usage{
get_netmhcpan_alleles(netmhcpan_folder_name = get_default_netmhcpan_folder())
}
\arguments{
\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\value{
a character vector with the NetMHCpan alleles
}
\description{
Get a list of the alleles supported by NetMHCpan
}
\examples{
if (is_netmhcpan_installed()) {
  get_netmhcpan_alleles()[1:5]
}
}
\author{
Richèl J.C. Bilderbeek
}
