% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhcpan_archive_filename.R
\name{get_netmhcpan_archive_filename}
\alias{get_netmhcpan_archive_filename}
\title{Get the filename of the NetMHCpan archive file}
\usage{
get_netmhcpan_archive_filename(
  netmhcpan_version = get_default_netmhcpan_version()
)
}
\arguments{
\item{netmhcpan_version}{the NetMHCpan version,
for example \code{4.1}.
Use \link{get_default_netmhcpan_version} to get the default
NetMHCpan version}
}
\value{
the filename of the NetMHCpan archive file
}
\description{
Get the filename of the NetMHCpan archive file
}
\examples{
get_netmhcpan_archive_filename()
}
