% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhc2pan_url.R
\name{get_netMHCpan_url}
\alias{get_netMHCpan_url}
\title{Deprecated, use \link{get_netMHCpan_bin_url} instead}
\usage{
get_netMHCpan_url()
}
\value{
a download URL
}
\description{
Deprecated, use \link{get_netMHCpan_bin_url} instead
}
\author{
Richèl J.C. Bilderbeek
}
