% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{install_netMHCpan}
\alias{install_netMHCpan}
\title{Install NetMHCpan}
\usage{
install_netMHCpan(
  netMHCpan_bin_url = get_netMHCpan_bin_url(),
  netMHCpan_data_url = get_netMHCpan_data_url(),
  verbose = FALSE,
  netMHCpan_archive_filename = get_netMHCpan_archive_filename(),
  netMHCpan_folder_name = get_default_netMHCpan_folder(),
  temp_local_file = tempfile(pattern = "netMHCpan_install_netMHCpan_")
)
}
\arguments{
\item{netMHCpan_bin_url}{URL to download the
NetMHCpan binary tarball file from,
similar to, for example,
\code{https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/data.tar.gz}.
Use \link{get_netMHCpan_bin_url} to get the default URL.}

\item{netMHCpan_data_url}{URL to download the
NetMHCpan data tarball file from,
similar to, for example,
\url{https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/data.tar.gz}.
Use \link{get_netmhc2pan_data_url} to get the default URL.}

\item{verbose}{set to TRUE for more output}

\item{netMHCpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netMHCpan_archive_filename} to get this filename.}

\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\value{
Nothing
}
\description{
Install NetMHCpan to a local folder, by downloading the binary and
data files. Use \link{install_netMHCpan_from_files} to install
NetMHCpan from files that are already downloaded.
}
\details{
These are three steps:

\itemize{
  \item Install the NetMHCpan binary, using \link{install_netMHCpan_bin}
  \item Install the NetMHCpan data, using \link{install_netMHCpan_data}
  \item Set up NetMHCpan, using \link{set_up_netMHCpan}
}
}
\examples{
\donttest{
  install_netMHCpan()
}
}
\author{
Richèl J.C. Bilderbeek
}
