% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhcpan_bin.R
\name{install_netmhcpan_bin}
\alias{install_netmhcpan_bin}
\title{Install the NetMHCpan binary to a local folder}
\usage{
install_netmhcpan_bin(
  netmhcpan_bin_url = get_netmhcpan_bin_url(),
  verbose = FALSE,
  netmhcpan_archive_filename = get_netmhcpan_archive_filename(),
  netmhcpan_folder_name = get_default_netmhcpan_folder(),
  temp_local_file = tempfile(pattern = "netmhcpan_install_netmhcpan_bin_")
)
}
\arguments{
\item{netmhcpan_bin_url}{URL to download the
NetMHCpan binary tarball file from,
similar to, for example,
\code{ FIX LINK}.
Use \link{get_netmhcpan_bin_url} to get the default URL.}

\item{verbose}{set to TRUE for more output}

\item{netmhcpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1b.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netmhcpan_archive_filename} to get this filename.}

\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\value{
Nothing
}
\description{
Install the NetMHCpan binary to a local folder.
}
\examples{
\donttest{
  install_netmhcpan_bin()
}
}
\author{
Richèl J.C. Bilderbeek
}
