% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan_bin_from_file.R
\name{install_netMHCpan_bin_from_file}
\alias{install_netMHCpan_bin_from_file}
\title{Install the NetMHCpan binary to a local folder}
\usage{
install_netMHCpan_bin_from_file(
  netMHCpan_bin_tarfile_path,
  verbose = FALSE,
  netMHCpan_archive_filename = get_netMHCpan_archive_filename(),
  netMHCpan_folder_name = get_default_netMHCpan_folder()
)
}
\arguments{
\item{netMHCpan_bin_tarfile_path}{path of the
  NetMHCpan binary tarball file
##This needs CHANGING}

\item{verbose}{set to TRUE for more output}

\item{netMHCpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netMHCpan_archive_filename} to get this filename.}

\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\value{
Nothing
}
\description{
Install the NetMHCpan binary to a local folder
}
\author{
Richèl J.C. Bilderbeek
}
