% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhcpan_from_files.R
\name{install_netmhcpan_from_files}
\alias{install_netmhcpan_from_files}
\title{Install NetMHCpan from files.}
\usage{
install_netmhcpan_from_files(
  netmhcpan_bin_tarfile_path,
  netmhcpan_data_tarfile_path,
  verbose = FALSE,
  netmhcpan_archive_filename = get_netmhcpan_archive_filename(),
  netmhcpan_folder_name = get_default_netmhcpan_folder(),
  temp_local_file = tempfile(pattern = "netmhcpan_install_netmhcpan_")
)
}
\arguments{
\item{netmhcpan_bin_tarfile_path}{path of the
NetMHCpan binary tarball file}

\item{netmhcpan_data_tarfile_path}{path of the
NetMHCpan data tarball file}

\item{verbose}{set to TRUE for more output}

\item{netmhcpan_archive_filename}{the NetMHCpan archive
filename, for example \code{netMHCpan-4.1b.Linux.tar.gz},
which is used in installation of NetMHCpan.
Use \link{get_netmhcpan_archive_filename} to get this filename.}

\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\value{
Nothing
}
\description{
Install NetMHCpan to a local folder from (already downloaded)
binary and date tarball files.
Use \link{install_netmhcpan} to install
NetMHCpan by downloading these files
}
\details{
These are three steps:

\itemize{
  \item Install the NetMHCpan binary,
    using \link{install_netmhcpan_bin_from_file}
  \item Install the NetMHCpan data,
    using \link{install_netmhcpan_data_from_file}
  \item Set up NetMHCpan, using \link{set_up_netmhcpan}
}
}
\author{
Richèl J.C. Bilderbeek
}
