% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhcpan_haplotype.R
\name{is_netmhcpan_haplotype}
\alias{is_netmhcpan_haplotype}
\title{Is the haplotype valid?}
\usage{
is_netmhcpan_haplotype(mhc_haplotype)
}
\arguments{
\item{mhc_haplotype}{one MHC haplotype, e.g. \code{DRB1_0101}.
See \link{get_netmhcpan_alleles} for a full list}
}
\value{
TRUE if the haplotype is a valid NetMHCpan haplotype
}
\description{
Determines if a haplotype has the same notation
as a NetMHCpan haplotype
}
\examples{
if (is_netmhcpan_installed()) {
  is_netmhcpan_haplotype("DRB1_0311")
  is_netmhcpan_haplotype("nonsense")
}
}
\author{
Richèl J.C. Bilderbeek
}
