% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhcpan_installed.R
\name{is_netmhcpan_installed}
\alias{is_netmhcpan_installed}
\title{Measure if NetMHCpan is installed locally}
\usage{
is_netmhcpan_installed(
  netmhcpan_folder_name = get_default_netmhcpan_folder(),
  verbose = FALSE
)
}
\arguments{
\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\value{
TRUE is NetMHCpan is installed locally,
  FALSE otherwise
}
\description{
Measure if NetMHCpan is installed locally
}
\author{
Richèl J.C. Bilderbeek
}
