% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan_report.R
\name{netMHCpan_report}
\alias{netMHCpan_report}
\title{Create a \code{netMHCpan} report, to be used when reporting bugs}
\usage{
netMHCpan_report(netMHCpan_folder_name = get_default_netMHCpan_folder())
}
\arguments{
\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\description{
Create a \code{netMHCpan} report, to be used when reporting bugs
}
\examples{
netMHCpan_report()
}
\author{
Richèl J.C. Bilderbeek
}
