% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmhc2pan_self_test.R
\name{netMHCpan_self_test}
\alias{netMHCpan_self_test}
\title{Test NetMHCpan}
\usage{
netMHCpan_self_test(
  netMHCpan_folder_name = get_default_netMHCpan_folder(),
  verbose = FALSE
)
}
\arguments{
\item{netMHCpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netMHCpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netMHCpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\description{
Test NetMHCpan
}
\author{
Richèl J.C. Bilderbeek
}
