% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_netmhcpan.R
\name{run_netmhcpan}
\alias{run_netmhcpan}
\title{Run NetMHCpan}
\usage{
run_netmhcpan(
  fasta_filename,
  alleles = "DRB1_0101",
  peptide_length = 15,
  netmhcpan_folder_name = get_default_netmhcpan_folder(),
  temp_xls_filename = netmhcpan::create_temp_xls_filename()
)
}
\arguments{
\item{fasta_filename}{the name of a FASTA file with protein sequences}

\item{alleles}{one or more alleles, e.g. \code{DRB1_0101}.
See \link{get_netmhcpan_alleles} for a full list}

\item{peptide_length}{length of a peptide}

\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}

\item{temp_xls_filename}{name for a temporary \code{xls} file,
which will be deleted automatically}
}
\value{
a data frame with the NetMHCpan results
}
\description{
Run NetMHCpan
}
\examples{
if (is_netmhcpan_installed()) {

  fasta_filename <- system.file(
    "extdata", "example.fasta", package = "netmhcpan"
  )
  run_netmhcpan(fasta_filename)

  # Two alleles
  alleles <- c("DRB1_0101", "DRB1_0102")
  # Run NetMHCpan with these two alleles
  run_netmhcpan(fasta_filename, alleles = alleles)
}
}
\author{
Richèl J.C. Bilderbeek
}
