% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_netmhcpan.R
\name{uninstall_netmhcpan}
\alias{uninstall_netmhcpan}
\title{Uninstall NetMHCpan}
\usage{
uninstall_netmhcpan(netmhcpan_folder_name = get_default_netmhcpan_folder())
}
\arguments{
\item{netmhcpan_folder_name}{the folder (to be) used by NetMHCpan.
From this location, a subfolder for NetMHCpan is created.
Use \link{get_default_netmhcpan_folder} to see the location
of the default NetMHCpan folder.
Use \link{get_default_netmhcpan_subfolder} to see the location
of the default NetMHCpan subfolder.}
}
\description{
Uninstall NetMHCpan
}
\author{
Richèl J.C. Bilderbeek
}
